requirements.txt
flask==2.3.3
pytesseract==0.3.10
pillow==10.0.0
PyMuPDF==1.23.8
PyPDF2==3.0.1
pdf2image==1.16.3
flask-sqlalchemy==3.0.3
flask-migrate==4.0.5
pix2tex
torch
torchvision